src/be/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
slowapi==0.1.9
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
ruff==0.1.6
black==23.11.0
pandas==2.1.3 
bcrypt==3.2.0 